% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_lat_long.R
\name{address_lat_long}
\alias{address_lat_long}
\title{Extracts the lat and long for each address}
\usage{
address_lat_long(data = df, address_column = "address")
}
\arguments{
\item{data}{dataframe from refine_authors}

\item{address_column}{name of column in quotes where the addresses are}
}
\description{
\code{address_lat_long} This function takes the author list output after the output has been synthesized for incorrect author matches. It contains a similarity score cutoff like read_authors. This however is to further constrain the list. New values ARE NOT created, instead it filters by the sim_score column in the output file. An output file is created using the 'root' argument that specifies the folder/file prefix for the output. The final file will be appended with '_final.csv'.
}
