% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\docType{data}
\name{countries}
\alias{countries}
\title{Names of all the countries in the world}
\format{a character vector of country names
\describe{
  \item{countries}{a character vector of country names}
  }}
\usage{
countries
}
\description{
#'
}
\keyword{datasets}
