% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_country_name.R
\name{extract_country_name}
\alias{extract_country_name}
\title{Extracts the country out of a refined author information}
\usage{
extract_country_name(data, address_column = "address")
}
\arguments{
\item{data}{dataframe from refine_authors}

\item{address_column}{name of column in quotes where the addresses are}
}
\description{
\code{extract_country_name} This function takes the author list output after the output has been synthesized for incorrect author matches. It contains a similarity score cutoff like read_authors. This however is to further constrain the list. New values ARE NOT created, instead it filters by the sim_score column in the output file. An output file is created using the 'root' argument that specifies the folder/file prefix for the output. The final file will be appended with '_final.csv'.
}
