% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_plot_coauthor_country.R
\name{net_plot_coauthor_country}
\alias{net_plot_coauthor_country}
\title{Creates a network diagram of coauthors' countries linked by reference, and with nodes arranged geographically}
\usage{
net_plot_coauthor_country(data, line_resolution = 10,
  mapRegion = "world")
}
\arguments{
\item{addresses}{the `address` element from the list outputted from the `georef_authors()`` function, containing geocoded address latitude and longitude locations.
#' @param authors_references output from the read_authors() function, which links author addresses together via AU_ID.}
}
\description{
\code{net_plot_coauthor_country} This function takes an addresses data.frame, links it to an authors_references dataset and plots a network diagram generated for countries of co-authorship.
}
