% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_addresses_country.R
\name{plot_addresses_country}
\alias{plot_addresses_country}
\title{Plot addresses, the number of which are summed by country_name}
\usage{
plot_addresses_country(data, mapRegion = "world")
}
\arguments{
\item{addresses}{the `address` element from the list outputted from the `georef_authors()`` function, containing geocoded address latitude and longitude locations.}
}
\description{
\code{plot_addresses_country} This function plots an addresses data.frame object by country name.
}
