% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_addresses_points.R
\name{plot_addresses_points}
\alias{plot_addresses_points}
\title{Plot address point locations on world map}
\usage{
plot_addresses_points(data, mapRegion = "world")
}
\arguments{
\item{addresses}{the `address` element from the list outputted from the `georef_authors()`` function, containing geocoded address latitude and longitude locations.}
}
\description{
\code{plot_addresses_points} This function plots an addresses data.frame object by point overlaid on the countries of the world.
}
