\name{refnet-package}
\alias{refnet-package}
\alias{refnet}
\docType{package}
\title{
Thomson Reuters Web of Knowledge and ISI Reference Data Tools
}
\description{
This function reads Thomson Reuters Web of Knowledge and ISI
format reference data files into an R friendly data format
and can optionally write the converted data to a friendly CSV
format.
}
\details{
\tabular{ll}{
Package: \tab refnet\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-02-06\cr
License: \tab GPL-3\cr
}
Package overview text will go here...
}
\author{
Forrest R. Stevens (University of Florida)

Maintainer: Forrest R. Stevens <forrest@ufl.edu>
}
\references{
Literature references will go here...

Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory
}
\keyword{ package }
\seealso{
Functions contained in this package:
\code{\link[refnet]{read_references}}, 
\code{\link[refnet]{read_authors}}, 
\code{\link[refnet]{remove_duplicates}}, 
\code{\link[refnet]{merge_records}}, 
\code{\link[refnet]{refnet_geocode}}, 
\code{\link[refnet]{refnet_geocode}}, 
\code{\link[refnet]{read_addresses}}, 
\code{\link[refnet]{plot_addresses_points}}, 
\code{\link[refnet]{plot_addresses_country}}, 
\code{\link[refnet]{net_plot_coauthor}}, 
\code{\link[refnet]{net_plot_coauthor_country}},
\code{\link[refnet]{net_plot_coauthor_address}}
}
\examples{
}