\name{plot_addresses_country}
\alias{plot_addresses_country}
\title{Plot addresses, the number of which are summed by country_name}
\usage{
plot_addresses_country(addresses)
}
\arguments{
  \item{addresses}{output from the read_addresses()
  function, containing geocoded address latitude and
  longitude locations.}
}
\description{
\code{plot_addresses_country} This function plots an
addresses data.frame object by country name.
}

